🗺️ Atlas: Extract database tests into sub-modules
