+++ .jules/atlas.md
@@ -18,3 +18,6 @@
 ## 2024-05-19 - Sub-modularizing the Persistent Engine Tests Blob
 **Tangle:** The `relvar-storage/src/persistent_engine/tests.rs` file had grown into a massive Blob containing nearly 1,400 lines of inline tests for various aspects of the persistent engine (transactions, recovery, garbage collection, and basic operations).
 **Blueprint:** Removed the `tests.rs` monolith and extracted its contents into a specialized sub-module structure under a new `relvar-storage/src/persistent_engine/tests/` directory. Created explicit sub-modules for `basic.rs`, `transaction.rs`, `recovery.rs`, and `gc.rs`, along with a `common.rs` for shared test initialization.
+## 2024-05-19 - Sub-modularizing the Database Tests Blob
+**Tangle:** The `relvar-core/src/database/tests.rs` file had grown into a 1,170 line Blob containing 45 inline tests. The God File mixed schema operations, DML, transaction boundaries, and virtual relvar tests.
+**Blueprint:** Removed the `tests.rs` monolith and extracted its contents into a specialized sub-module structure under a new `relvar-core/src/database/tests/` directory. Created explicit sub-modules for `schema.rs`, `data.rs`, `integrity.rs`, `transaction.rs`, and `virtual_relvar.rs`, matching the core database domain responsibilities, along with a `common.rs` for shared helpers.
