+++ relvar-core/src/database/tests/integrity.rs
@@ -1,132 +1,13 @@
+use super::common::*;
 use crate::constraints::ConstraintManagerError;
 use crate::constraints::expression::{CmpOp, ConstraintExpression, ValueOrRef};
 use crate::constraints::{CheckConstraint, CheckConstraints};
 use crate::database::Database;
 use crate::error::DatabaseError;
-use crate::storage_engine::{InMemoryEngine, StorageError};
+use crate::storage_engine::InMemoryEngine;
 use crate::tuple;
-use crate::types::RelationType;
-use crate::types::{ScalarType, TupleType};
-use crate::values::Relation;
+use crate::types::{RelationType, ScalarType, TupleType};
 use crate::values::ScalarValue;
-
-fn test_rel_type() -> RelationType {
-    RelationType::new(
-        TupleType::new()
-            .with_attribute("id", ScalarType::Int)
-            .with_attribute("name", ScalarType::String),
-    )
-}
-
-fn setup_parent_child_db() -> Database<InMemoryEngine> {
-    use crate::constraints::{KeyConstraints, PrimaryKey};
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    let parent_type = RelationType::new(
-        TupleType::new()
-            .with_attribute("id", ScalarType::Int)
-            .with_attribute("name", ScalarType::String),
-    );
-    let child_type = RelationType::new(
-        TupleType::new()
-            .with_attribute("child_id", ScalarType::Int)
-            .with_attribute("parent_id", ScalarType::Int),
-    );
-
-    db.create_relvar("PARENT", parent_type).unwrap();
-    db.create_relvar("CHILD", child_type).unwrap();
-
-    let pk = PrimaryKey::new(vec!["id".to_string()]).unwrap();
-    let parent_constraints = KeyConstraints::new().with_primary_key(pk);
-    db.set_key_constraints("PARENT", parent_constraints)
-        .unwrap();
-
-    db
-}
-
-#[test]
-fn test_create_and_query_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-    assert!(db.relvar_exists("TEST"));
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-}
-
-#[test]
-fn test_drop_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-    db.drop_relvar("TEST").unwrap();
-
-    assert!(!db.relvar_exists("TEST"));
-}
-
-#[test]
-fn test_delete() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-
-    let deleted = db
-        .delete("TEST", |t| t.get_typed::<i64>("id").unwrap() == 1)
-        .unwrap();
-    assert_eq!(deleted, 1);
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-}
-
-#[test]
-fn test_update() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    let updated = db
-        .update(
-            "TEST",
-            |t| t.get_typed::<i64>("id").unwrap() == 1,
-            |t| tuple! { id: t.get_typed::<i64>("id").unwrap(), name: "Alicia" },
-        )
-        .unwrap();
-    assert_eq!(updated, 1);
-
-    let result = db.query("TEST").unwrap();
-    let tuple = result.tuples().next().unwrap();
-    assert_eq!(tuple.get_typed::<String>("name").unwrap(), "Alicia");
-}
-
-#[test]
-fn test_transactions() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    db.begin().unwrap();
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 2);
-
-    db.rollback().unwrap();
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-}
 
 #[test]
 fn test_primary_key_constraint() {
@@ -293,32 +174,6 @@ fn test_type_constraint_positive_int() {
 }
 
 #[test]
-fn test_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-
-    // Define virtual relvar that projects just names
-    db.define_virtual_relvar(
-        "NAMES",
-        RelationType::new(TupleType::new().with_attribute("name", ScalarType::String)),
-        |db| {
-            let test = db.query("TEST")?;
-            Ok(test.project(&["name"]))
-        },
-    )
-    .unwrap();
-
-    // Query virtual relvar
-    let names = db.query("NAMES").unwrap();
-    assert_eq!(names.degree(), 1);
-    assert_eq!(names.cardinality(), 2);
-}
-
-#[test]
 fn test_delete_with_foreign_key_constraint() {
     use crate::constraints::{ForeignKey, ForeignKeyConstraints};
 
@@ -384,67 +239,6 @@ fn test_update_with_primary_key_violation() {
 }
 
 #[test]
-fn test_transaction_rollback_with_constraints() {
-    use crate::constraints::{KeyConstraints, PrimaryKey};
-
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    let pk = PrimaryKey::new(vec!["id".to_string()]).unwrap();
-    let constraints = KeyConstraints::new().with_primary_key(pk);
-    db.set_key_constraints("TEST", constraints).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    db.begin().unwrap();
-
-    // Make some changes
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-    db.delete("TEST", |t| t.get_typed::<i64>("id").unwrap() == 1)
-        .unwrap();
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-
-    // Rollback
-    db.rollback().unwrap();
-
-    // Should be back to original state
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-    assert!(result.contains(&tuple! { id: 1i64, name: "Alice" }));
-}
-
-#[test]
-fn test_error_relvar_not_found() {
-    let db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    let result = db.query("NONEXISTENT");
-    assert!(result.is_err());
-    // Error should be Storage(RelationNotFound) since it comes from the engine
-    match result {
-        Err(DatabaseError::Storage(StorageError::RelationNotFound(name))) => {
-            assert_eq!(name, "NONEXISTENT");
-        }
-        other => panic!("Expected RelationNotFound error, got: {:?}", other),
-    }
-}
-
-#[test]
-fn test_error_duplicate_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    let result = db.create_relvar("TEST", test_rel_type());
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::RelationAlreadyExists(_))
-    ));
-}
-
-#[test]
 fn test_set_type_constraints() {
     use crate::constraints::AttributeConstraints;
     use crate::constraints::type_constraint::TypeConstraint;
@@ -471,179 +265,6 @@ fn test_set_type_constraints() {
 
     // Invalid value should fail
     let result = db.insert("TEST", tuple! { id: 200i64, name: "Bob" });
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_transaction_commit() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    // Begin transaction
-    db.begin().unwrap();
-
-    // Make changes
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-
-    // Commit
-    db.commit().unwrap();
-
-    // Changes should persist
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 2);
-
-    // Transaction should no longer be active
-    assert!(!db.in_transaction);
-}
-
-#[test]
-fn test_commit_without_transaction() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    let result = db.commit();
-    assert!(result.is_err());
-    assert!(matches!(result, Err(DatabaseError::TransactionError(_))));
-}
-
-#[test]
-fn test_rollback_without_transaction() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    let result = db.rollback();
-    assert!(result.is_err());
-    assert!(matches!(result, Err(DatabaseError::TransactionError(_))));
-}
-
-#[test]
-fn test_begin_nested_transaction_fails() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.begin().unwrap();
-
-    // Try to begin another transaction
-    let result = db.begin();
-    assert!(result.is_err());
-    assert!(matches!(result, Err(DatabaseError::TransactionError(_))));
-}
-
-#[test]
-fn test_drop_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    // Define virtual relvar
-    db.define_virtual_relvar(
-        "VIRT",
-        RelationType::new(TupleType::new().with_attribute("name", ScalarType::String)),
-        |db| {
-            let test = db.query("TEST")?;
-            Ok(test.project(&["name"]))
-        },
-    )
-    .unwrap();
-
-    // Verify it exists
-    let result = db.query("VIRT");
-    assert!(result.is_ok());
-
-    // Drop it
-    db.drop_virtual_relvar("VIRT").unwrap();
-
-    // Should no longer exist
-    let result = db.query("VIRT");
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_drop_nonexistent_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    let result = db.drop_virtual_relvar("NONEXISTENT");
-    assert!(result.is_err());
-    assert!(matches!(result, Err(DatabaseError::RelationNotFound(_))));
-}
-
-#[test]
-fn test_cannot_insert_into_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.define_virtual_relvar(
-        "VIRT",
-        RelationType::new(TupleType::new().with_attribute("name", ScalarType::String)),
-        |db| {
-            let test = db.query("TEST")?;
-            Ok(test.project(&["name"]))
-        },
-    )
-    .unwrap();
-
-    // Try to insert
-    let result = db.insert("VIRT", tuple! { name: "Alice" });
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::CannotModifyVirtualRelvar(_))
-    ));
-}
-
-#[test]
-fn test_cannot_delete_from_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.define_virtual_relvar("VIRT", test_rel_type(), |db| db.query("TEST"))
-        .unwrap();
-
-    // Try to delete
-    let result = db.delete("VIRT", |_| true);
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::CannotModifyVirtualRelvar(_))
-    ));
-}
-
-#[test]
-fn test_cannot_update_virtual_relvar() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.define_virtual_relvar("VIRT", test_rel_type(), |db| db.query("TEST"))
-        .unwrap();
-
-    // Try to update
-    let result = db.update("VIRT", |_| true, |_| tuple! { id: 99i64, name: "X" });
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::CannotModifyVirtualRelvar(_))
-    ));
-}
-
-#[test]
-fn test_cannot_drop_base_relvar_when_virtual_depends_on_it() {
-    // This would require tracking dependencies, which might not be implemented
-    // Skipping for now as it may not be a current feature
-}
-
-#[test]
-fn test_virtual_relvar_error_propagation() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Define virtual relvar that queries nonexistent base
-    db.define_virtual_relvar("VIRT", test_rel_type(), |db| db.query("NONEXISTENT"))
-        .unwrap();
-
-    // Querying it should fail
-    let result = db.query("VIRT");
     assert!(result.is_err());
 }
 
@@ -749,72 +370,6 @@ fn test_delete_referenced_parent_success() {
             .unwrap(),
         "Parent1"
     );
-}
-
-#[test]
-fn test_delete_returns_count() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-    db.insert("TEST", tuple! { id: 3i64, name: "Charlie" })
-        .unwrap();
-
-    // Delete some tuples
-    let count = db
-        .delete("TEST", |t| t.get_typed::<i64>("id").unwrap() > 1)
-        .unwrap();
-    assert_eq!(count, 2);
-
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
-}
-
-#[test]
-fn test_update_returns_count() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-    db.insert("TEST", tuple! { id: 2i64, name: "Bob" }).unwrap();
-
-    // Update tuples
-    let count = db
-        .update(
-            "TEST",
-            |t| t.get_typed::<i64>("id").unwrap() == 1,
-            |_| {
-                tuple! { id: 1i64, name: "Alicia" }
-            },
-        )
-        .unwrap();
-
-    assert_eq!(count, 1);
-
-    let result = db.query("TEST").unwrap();
-    assert!(result.contains(&tuple! { id: 1i64, name: "Alicia" }));
-}
-
-#[test]
-fn test_delete_no_matches_returns_zero() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-    db.create_relvar("TEST", test_rel_type()).unwrap();
-
-    db.insert("TEST", tuple! { id: 1i64, name: "Alice" })
-        .unwrap();
-
-    // Delete with no matches
-    let count = db
-        .delete("TEST", |t| t.get_typed::<i64>("id").unwrap() > 100)
-        .unwrap();
-    assert_eq!(count, 0);
-
-    // Relation should be unchanged
-    let result = db.query("TEST").unwrap();
-    assert_eq!(result.cardinality(), 1);
 }
 
 #[test]
@@ -1033,138 +588,4 @@ fn test_update_constraint_violation_in_loop() {
             ConstraintManagerError::CheckConstraintViolation(_)
         ))
     ));
-}
-
-#[test]
-fn test_virtual_relvar_immutability_enforcement() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Create a log relation
-    let log_type = RelationType::new(TupleType::new().with_attribute("count", ScalarType::Int));
-    db.create_relvar("LOG", log_type).unwrap();
-
-    // Define a view. The compiler enforces that we cannot call mutable methods
-    // like insert() inside the evaluator because it receives &Database<E>, not &mut Database.
-    db.define_virtual_relvar("SAFE_VIEW", test_rel_type(), |db| {
-        // db.insert("LOG", ...); // This would cause compilation error!
-
-        // Read operations are allowed
-        let _ = db.query("LOG")?;
-
-        Ok(Relation::new(test_rel_type()))
-    })
-    .unwrap();
-
-    // Query the view
-    assert!(db.query("SAFE_VIEW").is_ok());
-}
-
-#[test]
-fn test_get_relvar_type() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Test base relvar
-    db.create_relvar("BASE", test_rel_type()).unwrap();
-    let base_type = db.get_relvar_type("BASE").unwrap();
-    assert_eq!(base_type, test_rel_type());
-
-    // Test virtual relvar
-    db.define_virtual_relvar("VIRTUAL", test_rel_type(), |db| db.query("BASE"))
-        .unwrap();
-    let virtual_type = db.get_relvar_type("VIRTUAL").unwrap();
-    assert_eq!(virtual_type, test_rel_type());
-
-    // Test missing relvar
-    let result = db.get_relvar_type("NONEXISTENT");
-    assert!(result.is_err());
-    assert!(matches!(result, Err(DatabaseError::Storage(_))));
-}
-
-#[test]
-fn test_transaction_errors() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Test commit when not in transaction
-    let result = db.commit();
-    assert!(result.is_err());
-    assert!(
-        matches!(result, Err(DatabaseError::TransactionError(msg)) if msg == "No transaction in progress")
-    );
-
-    // Test rollback when not in transaction
-    let result = db.rollback();
-    assert!(result.is_err());
-    assert!(
-        matches!(result, Err(DatabaseError::TransactionError(msg)) if msg == "No transaction in progress")
-    );
-
-    // Test nested begin
-    db.begin().unwrap();
-    let result = db.begin();
-    assert!(result.is_err());
-    assert!(
-        matches!(result, Err(DatabaseError::TransactionError(msg)) if msg == "Transaction already in progress")
-    );
-}
-
-#[test]
-fn test_define_virtual_relvar_already_exists() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Create base relvar
-    db.create_relvar("BASE", test_rel_type()).unwrap();
-
-    // Define virtual relvar
-    db.define_virtual_relvar("VIRTUAL", test_rel_type(), |db| db.query("BASE"))
-        .unwrap();
-
-    // Try to redefine virtual relvar
-    let result = db.define_virtual_relvar("VIRTUAL", test_rel_type(), |db| db.query("BASE"));
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::RelationAlreadyExists(_))
-    ));
-
-    // Try to define virtual relvar with same name as base relvar
-    let result = db.define_virtual_relvar("BASE", test_rel_type(), |db| db.query("BASE"));
-    assert!(result.is_err());
-    assert!(matches!(
-        result,
-        Err(DatabaseError::RelationAlreadyExists(_))
-    ));
-}
-
-#[test]
-fn test_list_relvars() {
-    let mut db: Database<InMemoryEngine> = Database::new(InMemoryEngine::new());
-
-    // Initially empty
-    assert!(db.list_relvars().is_empty());
-
-    let rel_type = test_rel_type();
-
-    // Add base relvars
-    db.create_relvar("TABLE_A", rel_type.clone()).unwrap();
-    db.create_relvar("TABLE_B", rel_type.clone()).unwrap();
-
-    let mut relvars = db.list_relvars();
-    relvars.sort();
-    assert_eq!(relvars, vec!["TABLE_A", "TABLE_B"]);
-
-    // Add virtual relvar
-    db.define_virtual_relvar("VIEW_C", rel_type.clone(), |_| {
-        Ok(Relation::new(test_rel_type()))
-    })
-    .unwrap();
-
-    let mut relvars2 = db.list_relvars();
-    relvars2.sort();
-    assert_eq!(relvars2, vec!["TABLE_A", "TABLE_B", "VIEW_C"]);
-
-    // Drop one
-    db.drop_relvar("TABLE_A").unwrap();
-    let mut relvars3 = db.list_relvars();
-    relvars3.sort();
-    assert_eq!(relvars3, vec!["TABLE_B", "VIEW_C"]);
 }
