+++ relvar-core/src/database/tests/mod.rs
@@ -1,0 +1,6 @@
+pub mod common;
+mod data;
+mod integrity;
+mod schema;
+mod transaction;
+mod virtual_relvar;
